requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
dynaconf[yaml]==3.1.4
reader==1.16
